test/pythonPG/requirements.txt
numpy==2.0.2
opencv-python==4.11.0.86
pyserial==3.5
